server/requirements.txt
fastapi==0.115.4
Flask==3.0.3
GPUtil==1.4.0
GPUtil==1.4.0
mysql_connector_repackaged==0.3.1
numpy==2.1.3
psutil==6.1.0
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9
starlette==0.41.2
torch==2.5.1
uvicorn==0.32.0
websockets==13.1
